requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic-settings==2.2.1
pydantic==2.6.3
requests==2.31.0
streamlit==1.32.0
pandas==2.2.1
chromadb==0.4.24
qdrant-client==1.8.0
langchain==0.1.11
langchain-community==0.0.27
langchain-openai==0.0.8
sentence-transformers==2.5.1
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.6.0
python-dotenv==1.0.1
torch --index-url https://download.pytorch.org/whl/cpu
